2.x: Subject/Processor improvements & small cleanup

+++ src/test/java/io/reactivex/single/SingleSubscribeTest.java
@@ -55,7 +55,7 @@ public class SingleSubscribeTest {
             }
         });
         
-        assertEquals((Integer)1, value[0]);
+        assertEquals(1, value[0]);
         assertNull(value[1]);
     }
